code/backend/services/cherry_blossom/requirements/requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
bcrypt==5.0.0
blinker==1.9.0
boto3==1.40.44
botocore==1.40.44
certifi==2025.8.3
# Editable Git install with no remote (cherry_blossom==0.0.0)
-e /Users/juanasoto/Desktop/JUMP/blossom/code/backend/services/cherry_blossom
click==8.3.0
distro==1.9.0
fastapi==0.118.0
Flask==3.1.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
lxml==6.0.2
MarkupSafe==3.0.3
myAws @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myAws
myDependencies @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myDependencies
myEncryption @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myEncryption
myExceptions @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myExceptions
myOpenai @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myOpenai
myOrm @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myOrm
openai==2.1.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
